Integrate: [ 24907] Silence warning in debug printf
[ 24937]
Silence some bcc32 compiler warnings

("Superfluous & with function")

[ 24938]
Silence some more bcc32 compiler warnings

("Suggest parentheses to clarify precedence")

[ 24939]
Silence yet more bcc32 compiler warnings

("Code has no effect")

[ 24944]
Silence another bcc32 compiler warning

("Possible use of '%s' before definition")

[ 24946]
More bcc32 compiler warnings to silence

("Suspicious pointer conversion")

[ 24947]
Silence bcc32 compiler warnings from win32/fcrypt.c (almost)

It still whines about 'in' and 'b' being assigned values that are
never used on lines 348 and 505 respectively (as side effects of the
calls to c2l and l2c respectively), but I'm not sure how to shut
them up.

[ 24950]
Oops! Change 24946 wasn't quite right

[ 24953]
Silence bcc32 compiler warnings following change 24945

("Possible use of '%s' before definition")

[ 24963]
Warning suppression.

[ 24977]
Fix typo in win32.c

Doesn't fix bcc32 + USE_LARGE_FILES, though :-(

[ 24994]
Silence compiler warning in SDBM_File

Unless I'm missing something it looks like "key" actually is used,
and the "key = key" assignment causes bcc32 to complain about use
before definition!

[ 24999]
Subject: [PATCH] Avoid redefinition warning for MinGW
Message-ID: <Mahogany-0.66.0-1232-20050627-220727.00@rbnet.it>
p4raw-link: @24999 on //depot/perl: eb9bdbdff170fbef354c89f26311bc9e4a0e7245
p4raw-link: @24994 on //depot/perl: aa0165afdda5e71e536a890f6b073619c221d9bd
p4raw-link: @24977 on //depot/perl: 56460430dcb7e8cada68e2a7dcfb1495b75076d1
p4raw-link: @24963 on //depot/perl: a882ab3181734afc6a4004c7cba75f148c724ba8
p4raw-link: @24953 on //depot/perl: 6300418df733fa6a39202abfea4908f73899f1ad
p4raw-link: @24950 on //depot/perl: d08fa7601ff56603f730a23e4363d5e951ce06c2
p4raw-link: @24947 on //depot/perl: 2534dd6e31cc0cf46c9012b51820cb53e717e775
p4raw-link: @24946 on //depot/perl: 8d3a710fc8101f114e42e18c7eae4598e4c21195
p4raw-link: @24944 on //depot/perl: 89e33a0587050e7ef2e88ba45c87444d8506f821
p4raw-link: @24939 on //depot/perl: 454f1e2628e3c3cf05341675e973e8df77c9b0ae
p4raw-link: @24938 on //depot/perl: 3b9e3074fddbae8590b8fbc95a2eac9fa7fc29f0
p4raw-link: @24937 on //depot/perl: 1c127fab650cf6427045f36d9e941565f5163fda
p4raw-link: @24907 on //depot/perl: cdd3ba141d0471cdda288698b2e3ab4be509f775

p4raw-id: //depot/maint-5.8/perl@25551
p4raw-integrated: from //depot/perl@25550 'copy in'
	ext/SDBM_File/SDBM_File.xs (@17750..) win32/fcrypt.c (@20702..)
	'merge in' win32/win32io.c (@24271..) perlvars.h (@24318..)
	win32/win32iop.h (@24974..)
p4raw-integrated: from //depot/perl@24953 'edit in' pp_ctl.c pp_sys.c
	(@24945..)
p4raw-integrated: from //depot/perl@24950 'edit in' doio.c (@24946..)
p4raw-integrated: from //depot/perl@24944 'merge in' toke.c (@24920..)
p4raw-integrated: from //depot/perl@24939 'merge in' pp_hot.c
	(@24871..)
p4raw-integrated: from //depot/perl@24938 'merge in' warnings.h
	warnings.pl (@24509..)
p4raw-integrated: from //depot/perl@24937 'edit in' win32/win32.c
	(@24883..) 'merge in' pp.c (@24871..)
p4raw-integrated: from //depot/perl@24907 'merge in' doop.c (@24871..)

+++ doop.c
@@ -1401,7 +1401,7 @@ Perl_do_kv(pTHX)
 		     hv_iterval(hv,entry) : avhv_iterval((AV*)hv,entry);
 	    DEBUG_H(Perl_sv_setpvf(aTHX_ tmpstr, "%lu%%%d=%lu",
 			    (unsigned long)HeHASH(entry),
-			    HvMAX(keys)+1,
+			    (int)HvMAX(keys)+1,
 			    (unsigned long)(HeHASH(entry) & HvMAX(keys))));
 	    SPAGAIN;
 	    XPUSHs(tmpstr);
